+++ ezcharts/__init__.py
@@ -1,6 +1,6 @@
 """Simple eCharts API."""
 
-__version__ = "0.10.1"
+__version__ = "0.10.2"
 
 import argparse
 import importlib